moving validations code to different classes
