+++ src/data/inventory.csv
@@ -1,3 +1,3 @@
-Clear Glass Bud Vases,50,149,Vases,2024-03-22
 Artificial Greenery Garlands 8ft,30,333,Greeneries,2024-02-28
 Rose Gold Sequin Tablecloths ,24,271,Table Runners,2024-02-28
+Clear Glass Bud Vases,50,149,Vases,2024-03-22
